Converted fetchProjectsActions.ts to typeScript.
