+++ app/types/allocationTypes.ts
@@ -27,23 +27,29 @@ export interface AllocationGridCell {
   id: string;
   project: string | null; // Project Name
   projectId: string | null;
-  projectCost: string | null;
-  projectCurrency: string | null;
-  projectStartDate: string | null;
-  projectEndDate: string | null;
-  projectLocation: string | null;
-  projectManager: string | null;
-  projectOvertimeAllowed: boolean | null;
-  projectSponsor: string | null;
-  projectStatus: string | null;
-  projectType: string | null;
+  projectCost?: number | null; // Only for Projects View
+  projectCurrency?: string | null; // Only for Projects View
+  projectStartDate?: string | null; // Only for Projects View
+  projectEndDate?: string | null; // Only for Projects View
+  projectLocation?: string | null; // Only for Projects View
+  projectManager?: string | null; // Only for Projects View
+  projectOvertimeAllowed?: boolean | null; // Only for Projects View
+  projectSponsor?: string | null; // Only for Projects View
+  projectStatus?: string | null; // Only for Projects View
+  projectType?: string | null; // Only for Projects View
   resource: string | null;
   resourceId: string | null;
   resourceType: string | null;
   role: string | null;
-  teamAllocationManager: string | null;
-  teamStatus: string | null;
+  teamAllocationManager?: string | null; // Only for Teams View
+  teamStatus?: string | null; // Only for Teams View
   teams: string | null;
   totalEffort: number | null;
-  [key: string]: AllocationGridCellData | string | number | boolean | null;
+  [key: string]:
+    | AllocationGridCellData
+    | string
+    | number
+    | boolean
+    | null
+    | undefined;
 }
